Moved fade to composite file.
